% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ADMM_EN_SMW.R
\name{ADMM_EN_SMW}
\alias{ADMM_EN_SMW}
\title{ADMM on l1 regularized quadratic program}
\usage{
ADMM_EN_SMW(Ainv, V, R, d, x0, lam, mu, maxits, tol, quiet, selector)
}
\arguments{
\item{Ainv}{Diagonal of \eqn{A^{-1}}{A^{-1}} term in SMW formula, where A is an n by n
positive definite coefficient matrix.}

\item{V}{Matrix from SMW formula.}

\item{R}{Upper triangular matrix in Cholesky decomposition of \eqn{I + UA^{-1}V}{I + U*Ainv*V}.}

\item{d}{nx1 dimensional column vector.}

\item{lam}{Regularization parameter for l1 penalty, must be greater than zero.}

\item{mu}{Augmented Lagrangian penalty parameter, must be greater than zero.}

\item{maxits}{Number of iterations to run}

\item{tol}{Vector of stopping tolerances, first value is absolute, second is relative tolerance.}

\item{quiet}{Logical controlling display of intermediate statistics.}

\item{selector}{Vector to choose which parameters in the discriminant vector will be used to calculate the
regularization terms. The size of the vector must be *p* the number of predictors. The
default value is a vector of all ones. This is currently only used for ordinal classification.}

\item{alpha}{Step length.}
}
\value{
\code{ADMM_EN_SMW} returns an object of \code{\link{class}} "\code{ADMM_EN_SMW}" including a list
with the following named components

\describe{
  \item{\code{call}}{The matched call.}
  \item{\code{x}}{Found solution.}
  \item{\code{y}}{Dual solution.}
  \item{\code{z}}{Slack variables.}
  \item{\code{k}}{Number of iterations used.}
}
}
\description{
Applies Alternating Direction Method of Multipliers to the l1-regularized quadratic program
\deqn{f(\mathbf{x}) + g(\mathbf{x}) = \frac{1}{2}\mathbf{x}^TA\mathbf{x} - d^T\mathbf{x} + \lambda |\mathbf{x}|_1}{f(x) + g(x) = 0.5*x^T*A*x - d^T*x + lambda*|x|_l1}
}
\details{
This function is used by other functions and should only be called explicitly for
debugging purposes.
}
\seealso{
Used by: \code{\link{SDAD}} and the \code{SDADcv} cross-validation version.
}
\keyword{internal}
