% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SZVD.R
\name{SZVD}
\alias{SZVD}
\alias{SZVD.default}
\title{Sparse Zero Variance Discriminant Analysis}
\usage{
SZVD(x, ...)

\method{SZVD}{default}(train, gamma, D, penalty, scaling, tol, maxits, beta,
  quiet)
}
\arguments{
\item{train}{Data matrix where first column is the response class.}

\item{gamma}{Regularization paramter controlling l1-penalty.}

\item{D}{Penalty dictionary basis matrix.}

\item{penalty}{Controls wether to apply reweighting of l1-penalty 
(using sigma = within-class std devs)}

\item{scaling}{Logical indicating whether to scale data such that each
feature has variance 1.}

\item{tol}{Stopping tolerances for ADMM algorithm, 
must include tol$rel and tol$abs.}

\item{maxits}{Maximum number of iterations used in the ADMM algorithm.}

\item{beta}{Parameter for augmented Lagrangian term in the ADMM algorithm.}
}
\value{
\code{SZVD} returns an object of \code{\link{class}} "\code{SZVD}" including a list
with the following named components

\describe{
  \item{\code{DVs}}{Discriminant vectors.}
  \item{\code{its}}{Number of iterations required to find DVs.}
  \item{\code{pen_scal}}{Weights used in reweighted l1-penalty.}
  \item{\code{N}}{Basis for the null-space of the sample within-class covariance.}
  \item{\code{w0}}{unpenalized zero-variance discriminants (initial solutions) plus B and W, etc.}
}

\code{NULL}
}
\description{
Applies SZVD heuristic for sparse zero-variance discriminant 
analysis to given training set.
}
\details{
This function will currently solve as a standalone function in accSDA for time comparison.
A wrapper function like ASDA will be created to use the functionality of plots and such.
Maybe call it ASZDA. For that purpose the individual ZVD function will need to be implemented.
}
\seealso{
Used by: \code{\link{SZVDcv}}.
}

