% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SZVDcv.R
\name{SZVDcv}
\alias{SZVDcv}
\alias{SZVDcv.default}
\title{Cross-validation of sparse zero variance discriminant analysis}
\usage{
SZVDcv(X, ...)

\method{SZVDcv}{default}(X, Y, folds, gams, beta, D, q, maxits, tol, ztol, feat,
  quiet)
}
\arguments{
\item{X}{n by p data matrix.}

\item{Y}{n by K indicator matrix.}

\item{folds}{number of folds to use in K-fold cross-validation.}

\item{gams}{Vector of regularization parameters for l1 penalty.
All elements must be greater than zero.}

\item{beta}{Augmented Lagrangian parameter. Must be greater than zero.}

\item{D}{Penalty dictionary basis matrix.}

\item{q}{Desired number of discriminant vectors.}

\item{maxits}{Number of iterations to runn ADMM algorithm.}

\item{tol}{Stopping tolerances for ADMM, must have tol$rel and tol$abs.}

\item{ztol}{Rounding tolerance for truncating entries to 0.}

\item{feat}{Maximum fraction of nonzero features desired in validation scheme.}

\item{quiet}{toggles between displaying intermediate statistics.}
}
\value{
\code{SZVDcv} returns an object of \code{\link{class}} "\code{SZVDcv}" 
       including a list with the named components \code{DVs} and \code{gambest}.
       Where \code{DVs} are the discriminant vectors for the best l1 regularization
       parameter and \code{gambest} is the best regularization parameter found
       in the cross-validation.

\code{NULL}
}
\description{
Applies alternating direction methods of multipliers to solve sparse
zero variance descriminant analysis.
}
\details{
This function will currently solve as a standalone function in accSDA for time comparison.
A wrapper function like ASDA will be created to use the functionality of plots and such.
Maybe call it ASZDA. For that purpose the individual ZVD function will need to be implemented.
}
\seealso{
Used by: \code{\link{SZVDcv}}.
}

