% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZVD.R
\name{ZVD}
\alias{ZVD}
\alias{ZVD.default}
\title{Zero Variance Discriminant Analysis}
\usage{
ZVD(A, ...)

\method{ZVD}{default}(A, scaling = FALSE, get_DVs = FALSE)
}
\arguments{
\item{A}{Matrix, where first column corresponds to class labels.}

\item{scaling}{Logical whether to rescale data so each feature has variance 1.}

\item{get_DVs}{Logical whether to obtain unpenalized zero-variance discriminant vectors.}
}
\value{
\code{SZVDcv} returns an object of \code{\link{class}} "\code{ZVD}"
       including a list with the following named components:
\describe{
  \item{\code{dvs}}{discriminant vectors (optional).}
  \item{\code{B}}{sample between-class covariance.}
  \item{\code{W}}{sample within-class covariance.}
  \item{\code{N}}{basis for the null space of the sample within-class covariance.}
  \item{\code{mu}}{training mean and variance scaling/centring terms}
  \item{\code{means}}{vectors of sample class-means.}
  \item{\code{k}}{number of classes in given data set.}
  \item{\code{labels}}{list of classes.}
  \item{\code{obs}}{matrix of data observations.}
  \item{\code{class_obs}}{Matrices of observations of each class.}
}

\code{NULL}
}
\description{
Implements the ZVD algorithm to solve dicriminant vectors.
}
\details{
This function should potentially be made internal for the release.
}
\seealso{
Used by: \code{\link{SZVDcv}}.
}
